Move EntityKey
