+++ snuba/datasets/cdc/groupassignee.py
@@ -1,5 +1,5 @@
 from snuba.datasets.dataset import Dataset
-from snuba.datasets.entities.factory import EntityKey
+from snuba.datasets.entities import EntityKey
 
 
 class GroupAssigneeDataset(Dataset):
