+++ snuba/datasets/outcomes_raw.py
@@ -1,5 +1,5 @@
 from snuba.datasets.dataset import Dataset
-from snuba.datasets.entities.factory import EntityKey
+from snuba.datasets.entities import EntityKey
 
 
 class OutcomesRawDataset(Dataset):
