+++ snuba/query/data_source/simple.py
@@ -1,6 +1,8 @@
 from dataclasses import dataclass
-from snuba.query.data_source import DataSource
+
 from snuba.clickhouse.columns import ColumnSet
+from snuba.datasets.entities import EntityKey
+from snuba.query.data_source import DataSource
 
 
 @dataclass(frozen=True)
@@ -9,7 +11,7 @@ class Entity(DataSource):
     Represents an Entity in the logical query.
     """
 
-    key: str
+    key: EntityKey
     schema: ColumnSet
 
     def get_columns(self) -> ColumnSet:
