+++ tests/query/processors/test_arrayjoin_optimizer.py
@@ -5,7 +5,8 @@ import pytest
 from snuba.clickhouse.astquery import AstSqlQuery
 from snuba.clickhouse.formatter import ClickhouseExpressionFormatter
 from snuba.clickhouse.query import Query as ClickhouseQuery
-from snuba.datasets.entities.factory import EntityKey, get_entity
+from snuba.datasets.entities import EntityKey
+from snuba.datasets.entities.factory import get_entity
 from snuba.datasets.factory import get_dataset
 from snuba.query.conditions import (
     BooleanFunctions,
